Translate the You Doubt minigame hints
* Translate the You Doubt minigame hints
There 50 hints. More than 40 of them have been verified in game. Maximum text length is 12 letters so some are even more vague than the original.

One odd thing about this game and these hints is that you are looking for the odd man out, meaning for instance that if the hint is about loving strawberries and two of the characters have strawberries mentioned in their bios, then the character that doesn't mention strawberries is the correct answer.

* Change the name of the "Sunset Carnival" game from "SUNSET FAIR" to "DUSKCARNIVAL". Shermie has a technique called "Shermie's Carnival" which she is imitating to pull vegetables. This game has nothing to do with an actual carnival or fair(e).